feat: Add TV list component
This commit adds a new TV list component with an HTML template and LESS file. The component displays a row of TV info cards using Angular's ngFor directive. The component also includes a TypeScript file that imports the TvShowResponse model and defines an input property for passing in an array of TV shows to display.