feat: Update MediaResponse id type to number
This commit updates the id property of the MediaResponse interface from a string to a number. This change was made to ensure consistency with other parts of the codebase that expect an id of type number. The tmdbId property was also removed as it is not being used in this context.