feat(media-info): Add play button for present movies
This commit adds a play button to the movie info page for present movies. It also adds a checkmark badge to the movie card component if the movie is present. The changes include modifications to HTML and TypeScript files, as well as adding new dependencies for badges and buttons from ng-zorro-antd library.