feat: Add TV view component
This commit adds a new TV view component that displays information about a selected TV show. The component includes an HTML template, LESS stylesheet, and TypeScript file. It imports the necessary dependencies and uses Angular's ActivatedRoute to retrieve the ID of the selected TV show from the URL parameters. It then calls MediaInfoService to fetch information about the TV show and displays it in the template.