feat: Add checkbox to filter only available media
This commit adds a new checkbox to the Trending page that allows users to filter media by availability. The `onlyAvailable` property is added to the component and used in API calls for movies and TV shows. When the checkbox is checked, it will trigger a call to `ngOnInit()` which will update the data displayed on the page based on user selection.