feat: Add media info component
This commit adds a new media info component that displays the name and poster of a movie or TV show. The component is now responsive and uses ng-zorro-antd components for styling. It also includes a service to fetch data from the server, which is used to display information about the selected media item.