feat: Add comment feature to movie and tv views
This commit adds a new CommentModule to the MediasModule, which allows users to add comments on movies and TV shows. The MovieViewComponent and TvViewComponent have been updated to display comments in a tab alongside media information. Pagination has also been added for displaying multiple comments. A new NewComment component has been created for adding new comments, which is displayed as an action button above the comment list.