feat: Add rating feature to movie view
This commit adds the ability for users to rate movies in the movie view component. The `NzRateModule` and `RatingModule` were imported into the `MediasModule`. The UI was updated to display ratings and allow users to add new ratings. The `RatingService` was added to retrieve user's rating and media's ratings from the backend API. Pagination was also added for displaying multiple ratings on a single page.