feat: Add user information display in account page
This commit adds a new card to the "Mes informations" tab in the auth-my-account component. The card displays user information such as username, email, first name, last name and status. The status is displayed using tags that indicate whether the user is active, subscribed or an administrator. Additionally, two new modules were imported to support this feature: NzDescriptionsModule and NzTagModule.