feat(user-info): add edit mode and form
This commit adds an edit mode to the user info component, allowing users with appropriate permissions to modify their own information or that of other users. A new form is displayed in edit mode, which allows for editing of username, email, first name and last name. The update process has also been implemented along with notifications for success/failure. Additionally, a badge color function was added to display different colors for each role. Finally, some dependencies were added to the user module: NzFormModule, NzSpaceModule and NzCheckboxModule.