feat: Update notifications to use success instead of info
This commit updates the notifications in both MovieInfoComponent and TvInfoComponent to use the success method instead of info. Additionally, it changes the text on a button in TvInfoComponent from "Suivre ce film" to "Suivre cette série". Finally, it fixes an issue with displaying percentages in StatisticsHistoryComponent by using the toFixed() method.