feat: Add image module and update watchlist UI
This commit adds the NzImageModule to the watchlist module, allowing images to be displayed in the UI. The watchlist component's HTML has been updated to display TV show and movie posters using this new module. Additionally, some minor changes have been made to improve the layout of the watchlist items. Finally, success notifications are now shown when a user updates an item's status in their watchlist.