feat: Add user and admin authorization to playlist component
This commit adds *ngIf directives to the playlist component's HTML template, allowing only authorized users (the owner of the playlist or an admin) to delete items from a playlist. It also adds two new properties to the component class, `user` and `isAdmin`, which are populated with data from the `AuthState` using Ngxs' `@Select` decorator. Finally, this commit updates the watchlist routing module by reordering navigation links and adding a constant for easy access to the playlists view link.