feat(media-info): Add watchlist functionality to TV show component
This commit adds the ability for users to add TV shows to their watchlist, change the status of a show on their watchlist, and remove a show from their watchlist. The changes include modifications to the HTML and TypeScript files of the TV info component, as well as importing necessary modules and services. The commit follows conventional commit methodology.