feat(movie-info): Add watchlist functionality
This commit adds the ability to add movies to a user's watchlist, change their status, and remove them. The movie-info component now displays a dropdown menu with options for adding or modifying the status of a movie in the user's watchlist. The changes include updating the HTML template and TypeScript file for this component, as well as importing new models and services.