refactor: Rename home.module.ts to homepage.module.ts
The commit renames the `home.module.ts` file to `homepage.module.ts`. It also updates the import statements in the `app-routing.module.ts`, and adds two new components, `app-noauth-home` and `app-auth-home`, to the `home.component.html` file. Finally, it modifies the `HomeComponent` class by adding a new property called `isAuth`, which is set based on whether or not a user is authenticated.