feat(homepage): add new UI design and features
This commit adds a new UI design to the homepage, including a carousel with images and descriptions of featured content, as well as categories for movies and TV shows. It also includes a login button that links to the login page. Additionally, this commit renames the `HomeComponent` to `HomePageComponent` for consistency with naming conventions. Finally, it imports several modules from ng-zorro-antd library such as NzLayoutModule, NzCarouselModule, NzDividerModule, NzGridModule, NzButtonModule, NzImageModule,NzSpaceModule,NzTypographyModule and NgOptimizedImage which are used in the new UI design.