feat: Add subscription check to watch buttons
This commit adds a subscription check to the watch buttons in the movie and episode info components. If the user is not subscribed, a tooltip will appear indicating that they must be subscribed to watch the content. The AuthState is used to determine if the user is subscribed or not.