refactor: Improve messaging UI and add friend list
This commit refactors the messaging component to improve its user interface. It adds a friend list on the left side of the screen, which allows users to select friends they want to chat with. The message list is now displayed in a more organized way using ng-zorro-antd components such as nz-list and nz-comment. Additionally, some CSS classes were modified or removed.

The sendMessage() method was also updated to prevent sending empty messages. Finally, this commit imports SharedModule, NzAvatarModule, and NzCommentModule modules that are used in the new implementation of the messaging component.