feat: Add disabled overlay to continue watching section
This commit adds a disabled overlay to the "continue watching" section of the home page. If the user is not subscribed, they will see an overlay with a message indicating that they need to subscribe in order to continue watching. The overlay also disables clicks on the component. Additionally, this commit adds the NzToolTipModule to HomeModule and removes commented-out code from AuthHomeComponent's HTML file.