feat: Add mediaType input to rating-list component
This commit adds a new input called mediaType to the rating-list component. The mediaType input is used in the HTML template to display either movie or TV show names based on the type of media being rated. This change improves the readability and accuracy of the displayed information for users.