feat: Add media type to MediaLinkPipe and MediaNamePipe
The MediaLinkPipe and MediaNamePipe now accept a media type parameter, which can be either 'movie' or 'tv'. The pipes will return the appropriate link or name based on the given media ID and type. Additionally, the MovieNamePipe has been renamed to MediaNamePipe to reflect its new functionality. The TvNamePipe has been removed since it is no longer needed with these changes.