feat: add delete user functionality
This commit adds the ability to delete a user account by adding a new `delete()` method to the `UserService` class. Additionally, an `adminDeleteUser()` method was added for administrative users to delete other users' accounts. The `tap` operator was also imported from RxJS and the `delay` operator was added as well.