feat: Add delete user functionality
This commit adds the ability to delete a user account. A modal is displayed when an admin clicks on the "Supprimer ce compte" button, which allows them to confirm or cancel the deletion. The modal displays different results based on whether the deletion was successful or not. Additionally, this commit includes changes to import NzModalModule and NzResultModule in social-network.module.ts and adds isAdmin$ observable in user-view.component.ts.