feat: Add file-stats component
This commit adds a new component called "file-stats" which displays general information about media files. The component includes a card with statistics on the number of media files, disk space used and available, and a progress bar showing the percentage of available disk space. It also includes a refresh button to update the data displayed. The component is implemented using Angular and uses MediaFileService to retrieve data from the server.