feat: Add responsive design to movie-info-card component
This commit adds a responsive design to the movie-info-card component by changing the size of the card based on whether it is being viewed on a phone or desktop. It also hides certain elements on mobile devices and displays them only on desktops. The changes were made in both the HTML and TypeScript files, with the addition of a new property 'isOnPhone' in TypeScript file that detects if user is viewing from phone or not using BreakpointObserver.