feat: Add responsive design for mobile devices
This commit adds responsive design for mobile devices by adjusting the height, padding, and font size of various elements. It also adds a new variable `isOnPhone` to detect if the user is on a phone or not using `BreakpointObserver`. Finally, it hides certain menu items on mobile devices using `[class]` binding.