feat: Add responsive design to Trending page
This commit adds a responsive design to the Trending page by using the Angular CDK layout module. The `isOnPhone` variable is used to determine whether the user is on a phone or not, and adjusts the size of certain elements accordingly. Pagination and tabset components now have their sizes adjusted based on this variable as well.