feat(watchlist): Add responsive design to watchlist component
The watchlist component now has a responsive design that adjusts the size of tabs and input fields based on screen size. This was achieved by adding a new variable `isOnPhone` which is set to true when the screen size matches that of a phone. The `nzSize` property was also added to adjust the size of some elements in the template. A new import statement for `BreakpointObserver` was added to the TypeScript file, and an observable was created in the constructor to listen for changes in screen size.