feat: Add search bar and hide delete button on mobile view
This commit adds a search bar to the playlist component for both movies and episodes. It also hides the delete button on mobile view by adding a class "hide-on-phone" to it. The same class is added to the input-group of the search bar so that it is hidden on desktop view.