feat: Add responsive design to statistics history component
The changes include adding a new variable `isOnPhone` and using it to conditionally display elements based on the screen size. The breakpoint observer is used to detect when the user is on a phone and set the value of `isOnPhone`. The delete button now has an icon and its text is hidden on phones.