feat: Add responsive design to user-info component
This commit adds a responsive design to the user-info component by using the Angular CDK layout module. The changes include adding a hide-on-desktop and hide-on-phone class, which hides certain elements based on screen size. Additionally, the isOnPhone property is added to determine if the user is on a phone or not. This property is used to adjust the size and layout of some elements in the component.