feat: Add friend list component actions with titles
This commit adds titles to the buttons in the friend list component. The titles are "Accepter la demande d'ami", "Refuser la demande d'ami", "Bloquer l'utilisateur" and "Supprimer la demande d'ami". Additionally, this commit adds a new property `isOnPhone` to the FriendListComponent class which is set based on whether or not the user is on a phone screen.