feat(messaging): add responsive design for mobile devices
The messaging component now has a responsive design that adjusts to the screen size of mobile devices. The user list and message list have been modified to fit smaller screens, and the `isOnPhone` variable has been added to detect when the user is on a phone. This was achieved by adding `[nzXs]` property to some elements in the HTML file, modifying CSS styles in the LESS file, and using `BreakpointObserver` from Angular CDK Layout library in TypeScript.