feat: Add responsive design to playlist stream component
This commit adds a responsive design to the playlist stream component by adjusting the size of the card and list items based on screen size. The height of the container is also adjusted for better scrolling experience. The changes were made in HTML, LESS, and TypeScript files. A new variable `isOnPhone` was added to detect if the user is on a phone or not using Angular's `BreakpointObserver`.