feat: Improve movie-by-genre component layout and responsiveness
This commit adds a new h4 tag to replace the nz-page-header element, which was commented out. It also adjusts the size of the nz-card and nz-pagination elements based on whether the user is on a phone or not. Finally, it introduces a new variable called "isOnPhone" that is set to true if the user is on a phone-sized screen.