feat: Improve TV by Network component UI
This commit improves the user interface of the TV by Network component. The page header has been commented out and replaced with a smaller heading, and pagination and card size have been adjusted based on screen size. Additionally, a new variable `isOnPhone` has been added to detect if the user is on a phone or not.