feat(subscription): add conditional rendering to Actions section
This commit adds a condition to the Actions section of the subscription component's HTML file. The "Actions" title will only be displayed if the user is on a phone. Additionally, the cancel subscription button has been modified to only display a pop-up confirmation message instead of calling the cancelSubscription() function directly.