feat: Add new methods to retrieve short info for movies and TV shows
This commit adds two new methods to the MediaInfoService that allow retrieving short information about multiple movies or TV shows at once. The methods use the TMDb IDs of the media items as input and return an array of MovieResponse or TvShowResponse objects respectively. The implementation uses HttpParams to pass the IDs as query parameters in the GET request.