feat: Add new methods to MediaInfoService
This commit adds two new methods to the MediaInfoService class that allow for retrieving all episodes of a TV show and their IDs. It also modifies existing methods to use HTTP POST instead of GET requests. Additionally, it includes changes to the StatsShowDailyViewedGenreComponent that only updates data if there are any episode stats available and adds a tap operator for debugging purposes.