feat: Add episode list to watchlist component
This commit adds an episode list to the watchlist component. The episode list is now displayed in a collapsible panel and includes a dropdown menu for changing the status of each episode. Additionally, the height of the viewport has been increased from 400px to 450px. The width of the .episode-list class has been commented out as it is no longer needed.